Add new layer to remove modtaps from certain keys
macOS has a handy way of typing diacritics which you can invoke by
holding a key and pressing the number corresponding to the diacritic
version of the key you want to type. Mod-taps intercept holds and for
that reason this handy diacritic popup cannot be invoked. A new layer is
created for when diacritic versions of some mod-tap keys are needed,
which can be activated by a toggle or a hold.