Change symbol layer fake LT key processing
During fast typing, the fake layer tap keys in the symbol layer
sometimes printed the keycode in the base layer because the records
arrived in the wrong order to `process_record_user`. The tap keycodes
are now processed in `pre_process_record_user` and later if the press is
resolved to a hold, the tap keycode is deleted and the hold action is
performed instead.

Also the code related to the one-shot handling of the symbol layer is
removed as now the symbol layer key is a `MO` key.